Break up character.go
